Update Event Form Display and Positioning
Updated the event form to be displayed by default and added absolute positioning to it.
This allows the form to appear at the location of the click event when a date is selected.
Also, added new CSS styles for the event form and
updated the selectDate function in script.js to
include the click event as a parameter.